requirements.txt
google-generativeai
requests
dotenv-python